rename: 로그인, 회원가입 api 관련 파일 명을 account로 변경
user.ts -> account.ts
