+++ src/hooks/signUp/mutate/usePostCoachSignUp.ts
@@ -1,5 +1,5 @@
 import { useMutation } from "@tanstack/react-query";
-import { postSignUp } from "~/api/user";
+import { postSignUp } from "~/api/account";
 
 interface Props {
   // userType: "external" | "player" | "coach";
